Issue #50 (Rework Makefile's mkdir mechanism to avoid issues in parallel builds): implemented implicitly creating the necessary directories for any target placed in a particular folder by introducing a "dummy" file representing the directory in questions. Removed Makefile.inc in sample folders by relying on DEPDIR's Makefile.inc. Include Makefile.inc at the bottom of each Makefile rather than in between (necessary since Makefile.inc now also contains a rule which cannot be the default rule; also this prepares for moving pattern rules into Makefile.inc).
As a general cleanup, remove the rule(s) in the NEK sample which are installing into DEPDIR. Really this is an awful solution where the sample code installs into the library's directory structure (and this cannot be preserved). Any NEK-related code can still do it the other way around and simply rely on the sample folder. In another cleanup stage, one could also remove the sample related rules in LIBXSMM's Makefile doing various stuff (testing, script generation, etc.). Really this was never intended to be a solution for dealing with Travis (and there are better ways to do this).