Introduced Makefile state tracking such that a rebuild is triggered whenever an influential environment variable changed or when the user requested a different build set. This is to ensure multiple version of libxsmm (mainly STATIC vs. non-STATIC) are built separately, and correctly considering certain defines (e.g., __STATIC). This commit is to better support maintainer building packages for a Linux distribution.
The Makefile.inc supplies a file target called ".state". This can be picked by any client-Makefile by adding a rule at the end of the Makefile which always triggers (.PHONY) the update of ".state". The root of all dependencies in the Makefile might be depending on the ".state" file. The latter is only updated if the internal Makefile state changed from build to build.