fix several issue: - binary was incorrectly enabled for stochastic rounding -> it didn't   use stochastic rounding at all - replicate_cal_var was incorrectly enabled for stochstic roudning -> it   didn't use stochastic rounding at all - the C++ roudning sequence has been updated/algined with other using   nested ternart assigment - the instruction sequence used in the code gen was buggy and inefficient (two many   instructions), it has been rewritten from scratch
With these changes the current unit test (with prng state set to 0)
pass. This is all unary simple and binary simple. ReLU, transforms, etc.
are not still not enabled. It need to be determined which TPP should
offer support for stochatic rounding when the output datatype is BF8.

Also for the current state, allow list for ARCH other than AVX512 need
to be updated.