Enable exhaustive x86 TPP testing (#832)
a) added all archs we support on x86, there are a handful of bugfixes in this PR to fix broken arch decision for chip we don't have in our hands (SNB, etc.)
b) for layernorm I ran an exhaustive 100x100x100 grid under GCC and made CI compiler GCC. the largest error was "0.006932419423397524933794", so I set the CI boundary to 0.007, I have full logs from this run if you guys have any better idea.
c) the GEMM test were the toughest nut to crack as they had really bad norms from time to time after playing with several strategies (and none of them led anywhere as we want to freedom to reorder "K" in the our GEMM kernels (we might want to add an env variable to switch that feature off, I will create an issue). So, I settled for the following fix:
I) when the check norm is outrageously high (means close to 1), we look at the eltwise inf-norm, if this is in check we use the inf-norm instead of the check-norm. This is normally the case for M=1-3, N=1-3 and K>60, so accumulation chain related.
II) when this new norm from I) is bigger than 0.009 we assume an accumulation reorder issue. In this case we replay the current test with matrix A set to identity matrix. B and C will remain fully random. When now the norm is still bigger than 0.009 we flag the test as failure, if it was acc-related it should be 0.0 and therefore the test automatically passes and we move on to the next test case.